pyproject.toml
[project]
name = "telegram-request-bot"
version = "0.1.0"
description = "Telegram channel request manager"
authors = [
    {name = "SamsonHyacinth",email = "samsonsanou@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot (>=22.0,<23.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
